+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/ActivityHelper.cs
@@ -34,7 +34,7 @@ internal static class ActivityHelper
     internal static readonly object StartedButNotSampledObj = new();
 
     private const string BaggageSlotName = "otel.baggage";
-    private static readonly Func<HttpRequest, string, IEnumerable<string>> HttpRequestHeaderValuesGetter = (request, name) => request.Headers.GetValues(name);
+    private static readonly Func<HttpRequestBase, string, IEnumerable<string>> HttpRequestHeaderValuesGetter = (request, name) => request.Headers.GetValues(name);
     private static readonly ActivitySource AspNetSource = new(
         AspNetSourceName,
         typeof(ActivityHelper).Assembly.GetPackageVersion());
@@ -44,14 +44,14 @@ internal static class ActivityHelper
 
     /// <summary>
     /// Try to get the started <see cref="Activity"/> for the running <see
-    /// cref="HttpContext"/>.
+    /// cref="HttpContextBase"/>.
     /// </summary>
-    /// <param name="context"><see cref="HttpContext"/>.</param>
+    /// <param name="context"><see cref="HttpContextBase"/>.</param>
     /// <param name="aspNetActivity">Started <see cref="Activity"/> or <see
     /// langword="null"/> if 1) start has not been called or 2) start was
     /// called but sampling decided not to create an instance.</param>
     /// <returns><see langword="true"/> if start has been called.</returns>
-    public static bool HasStarted(HttpContext context, out Activity? aspNetActivity)
+    public static bool HasStarted(HttpContextBase context, out Activity? aspNetActivity)
     {
         var itemValue = context.Items[ContextKey];
         if (itemValue is ContextHolder contextHolder)
@@ -68,10 +68,10 @@ internal static class ActivityHelper
     /// Creates root (first level) activity that describes incoming request.
     /// </summary>
     /// <param name="textMapPropagator"><see cref="TextMapPropagator"/>.</param>
-    /// <param name="context"><see cref="HttpContext"/>.</param>
+    /// <param name="context"><see cref="HttpContextBase"/>.</param>
     /// <param name="onRequestStartedCallback">Callback action.</param>
     /// <returns>New root activity.</returns>
-    public static Activity? StartAspNetActivity(TextMapPropagator textMapPropagator, HttpContext context, Action<Activity?, HttpContext>? onRequestStartedCallback)
+    public static Activity? StartAspNetActivity(TextMapPropagator textMapPropagator, HttpContextBase context, Action<Activity?, HttpContextBase>? onRequestStartedCallback)
     {
         var propagationContext = textMapPropagator.Extract(default, context.Request, HttpRequestHeaderValuesGetter);
 
@@ -132,10 +132,10 @@ internal static class ActivityHelper
     /// </summary>
     /// <param name="textMapPropagator"><see cref="TextMapPropagator"/>.</param>
     /// <param name="aspNetActivity"><see cref="Activity"/>.</param>
-    /// <param name="context"><see cref="HttpContext"/>.</param>
+    /// <param name="context"><see cref="HttpContextBase"/>.</param>
     /// <param name="onRequestStoppedCallback">Callback action.</param>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void StopAspNetActivity(TextMapPropagator textMapPropagator, Activity? aspNetActivity, HttpContext context, Action<Activity?, HttpContext>? onRequestStoppedCallback)
+    public static void StopAspNetActivity(TextMapPropagator textMapPropagator, Activity? aspNetActivity, HttpContextBase context, Action<Activity?, HttpContextBase>? onRequestStoppedCallback)
     {
         if (aspNetActivity == null)
         {
@@ -184,14 +184,14 @@ internal static class ActivityHelper
     }
 
     /// <summary>
-    /// Notifies listeners about an unhandled exception thrown on the <see cref="HttpContext"/>.
+    /// Notifies listeners about an unhandled exception thrown on the <see cref="HttpContextBase"/>.
     /// </summary>
     /// <param name="aspNetActivity"><see cref="Activity"/>.</param>
-    /// <param name="context"><see cref="HttpContext"/>.</param>
+    /// <param name="context"><see cref="HttpContextBase"/>.</param>
     /// <param name="exception"><see cref="Exception"/>.</param>
     /// <param name="onExceptionCallback">Callback action.</param>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void WriteActivityException(Activity? aspNetActivity, HttpContext context, Exception exception, Action<Activity?, HttpContext, Exception>? onExceptionCallback)
+    public static void WriteActivityException(Activity? aspNetActivity, HttpContextBase context, Exception exception, Action<Activity?, HttpContextBase, Exception>? onExceptionCallback)
     {
         try
         {
@@ -211,9 +211,9 @@ internal static class ActivityHelper
     /// This method is intended to restore the current activity in order to correlate the child
     /// activities with the root activity of the request.
     /// </summary>
-    /// <param name="context"><see cref="HttpContext"/>.</param>
+    /// <param name="context"><see cref="HttpContextBase"/>.</param>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static void RestoreContextIfNeeded(HttpContext context)
+    internal static void RestoreContextIfNeeded(HttpContextBase context)
     {
         if (context.Items[ContextKey] is ContextHolder contextHolder && Activity.Current != contextHolder.Activity)
         {