+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetTraceInstrumentationOptions.cs
@@ -50,7 +50,7 @@ public class AspNetTraceInstrumentationOptions
     /// exception the request is filtered out (NOT collected).</item>
     /// </list>
     /// </remarks>
-    public Func<HttpContext, bool>? Filter { get; set; }
+    public Func<HttpContextBase, bool>? Filter { get; set; }
 
     /// <summary>
     /// Gets or sets an action to enrich an Activity.
@@ -59,7 +59,7 @@ public class AspNetTraceInstrumentationOptions
     /// <para><see cref="Activity"/>: the activity being enriched.</para>
     /// <para><see cref="HttpRequest"/>: the HttpRequest object from which additional information can be extracted to enrich the activity.</para>
     /// </remarks>
-    public Action<Activity, HttpRequest>? EnrichWithHttpRequest { get; set; }
+    public Action<Activity, HttpRequestBase>? EnrichWithHttpRequest { get; set; }
 
     /// <summary>
     /// Gets or sets an action to enrich an Activity.
@@ -68,7 +68,7 @@ public class AspNetTraceInstrumentationOptions
     /// <para><see cref="Activity"/>: the activity being enriched.</para>
     /// <para><see cref="HttpResponse"/>: the HttpResponse object from which additional information can be extracted to enrich the activity.</para>
     /// </remarks>
-    public Action<Activity, HttpResponse>? EnrichWithHttpResponse { get; set; }
+    public Action<Activity, HttpResponseBase>? EnrichWithHttpResponse { get; set; }
 
     /// <summary>
     /// Gets or sets an action to enrich an Activity.